Move type methods into instanceImpl
