+++ grakn-graph/src/main/java/ai/grakn/graph/internal/TypeImpl.java
@@ -58,7 +58,7 @@ import java.util.stream.Collectors;
  */
 class TypeImpl<T extends Type, V extends Instance> extends ConceptImpl<T, Type> implements Type {
     TypeImpl(AbstractGraknGraph graknGraph, Vertex v, Optional<T> superType, Optional<Boolean> isImplicit) {
-        super(graknGraph, v, Optional.empty());
+        super(graknGraph, v);
         superType.ifPresent(this::superType);
         isImplicit.ifPresent(i -> setImmutableProperty(Schema.ConceptProperty.IS_IMPLICIT, i, getProperty(Schema.ConceptProperty.IS_IMPLICIT), Function.identity()));
     }
